+++ src/Akeneo/Pim/Structure/Component/Normalizer/Versioning/AttributeNormalizer.php
@@ -3,8 +3,10 @@
 namespace Akeneo\Pim\Structure\Component\Normalizer\Versioning;
 
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
+use Doctrine\Common\Collections\ArrayCollection;
 use Symfony\Component\Serializer\Normalizer\CacheableSupportsMethodInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+use Webmozart\Assert\Assert;
 
 /**
  * A normalizer to transform an AttributeInterface entity into a flat array
@@ -83,6 +85,7 @@ class AttributeNormalizer implements NormalizerInterface, CacheableSupportsMetho
     protected function normalizeOptions(AttributeInterface $attribute): ?string
     {
         $options = $attribute->getOptions();
+        Assert::isInstanceOf($options, ArrayCollection::class);
         if ($options->isEmpty()) {
             return null;
         }